Split string helpers into separate submodule
